backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.10.4
click==8.3.0
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
modbus==3.2
numpy==2.2.6
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
pymodbus==3.11.3
pyserial==3.5
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
websockets==15.0.1
pywebview==5.3
pyinstaller==6.11.1
reportlab==4.2.5
